Rebase + moved verification to utxo-ledger-flow
